test: split e2e test file
